Factored table editor functions into their own file
